feat: restructure app home functionality with customizations and actions integration
