+++ server/slack/actions/index.ts
@@ -1,0 +1,3 @@
+import { customizations } from '../features/customizations';
+
+export const actions = [...customizations.actions];
