+++ server/slack/app.ts
@@ -3,19 +3,31 @@ import { env } from '~/env';
 import { buildCache } from '~/lib/allowed-users';
 import logger from '~/lib/logger';
 import type { SlackApp } from '~/types';
+import { actions } from './actions';
 import { events } from './events';
 import { register as registerAppHomeOpened } from './events/app-home-opened';
 import { register as registerAssistantThreadContextChanged } from './events/assistant-thread-context-changed';
 import { register as registerAssistantThreadStarted } from './events/assistant-thread-started';
+import { views } from './views';
 
 function registerApp(app: App) {
   buildCache(app);
+
   for (const event of events) {
     app.event(event.name, event.execute);
   }
+
   registerAssistantThreadStarted(app);
   registerAssistantThreadContextChanged(app);
   registerAppHomeOpened(app);
+
+  for (const action of actions) {
+    app.action(action.name, action.execute);
+  }
+
+  for (const view of views) {
+    app.view(view.name, view.execute);
+  }
 }
 
 export function createSlackApp(): SlackApp {
