--- server/slack/events/app-home-opened/on-view.ts
@@ -1,47 +1,0 @@
-import type {
-  AllMiddlewareArgs,
-  App,
-  SlackViewMiddlewareArgs,
-  ViewSubmitAction,
-} from '@slack/bolt';
-import {
-  clearUserCustomization,
-  setUserCustomization,
-} from '~/db/queries/customizations';
-import logger from '~/lib/logger';
-import { toLogError } from '~/utils/error';
-import { asRecord } from '~/utils/record';
-
-type ViewArgs = SlackViewMiddlewareArgs<ViewSubmitAction> & AllMiddlewareArgs;
-
-export function registerViews(
-  app: App,
-  publishHome: (client: ViewArgs['client'], userId: string) => Promise<void>
-): void {
-  app.view(
-    'home_save_prompt',
-    async ({ ack, view, body, client }: ViewArgs) => {
-      await ack();
-      const userId = body.user.id;
-      const promptInput = asRecord(
-        view.state.values.prompt_block?.prompt_input
-      )?.value;
-      const prompt = (
-        typeof promptInput === 'string' ? promptInput : ''
-      ).trim();
-      try {
-        if (prompt) {
-          await setUserCustomization(userId, { prompt });
-        } else {
-          await clearUserCustomization(userId);
-        }
-        await publishHome(client, userId);
-      } catch (error) {
-        logger.warn(
-          { ...toLogError(error), userId },
-          'Failed to save custom prompt'
-        );
-      }
-    }
-  );
-}
