+++ server/slack/events/app-home-opened/view.ts
@@ -4,13 +4,9 @@ import {
   Blocks,
   Elements,
   HomeTab,
-  Modal,
   setIfTruthy,
 } from 'slack-block-builder';
-import type {
-  SlackHomeTabDto,
-  SlackModalDto,
-} from 'slack-block-builder/dist/internal';
+import type { SlackHomeTabDto } from 'slack-block-builder/dist/internal';
 import { appHome } from '~/config';
 import type { ScheduledTask } from '~/db/schema';
 import { personas } from '~/lib/ai/prompts/chat/presets';
@@ -140,34 +136,6 @@ export function buildHomeView(
         )
       ),
       tasks.map((task) => buildTaskBlock(task))
-    )
-    .buildToObject();
-}
-
-export function buildPromptModal(currentPrompt: string | null): SlackModalDto {
-  return Modal({
-    title: 'Custom Instructions',
-    submit: 'Save',
-    close: 'Cancel',
-    callbackId: 'home_save_prompt',
-  })
-    .blocks(
-      Blocks.Section({
-        text: 'Tell Gorkie how you want it to behave in every conversation - your preferred language, tone, name, or anything else.',
-      }),
-      Blocks.Input({
-        blockId: 'prompt_block',
-        label: 'Your instructions',
-      }).element(
-        Elements.TextInput({
-          actionId: 'prompt_input',
-          multiline: true,
-          maxLength: 3000,
-          placeholder:
-            'e.g. Always reply in Spanish. Keep responses concise. My name is Alex.',
-          initialValue: currentPrompt ?? undefined,
-        })
-      )
     )
     .buildToObject();
 }
