+++ server/slack/features/customizations/index.ts
@@ -1,0 +1,6 @@
+import { prompts } from './prompts';
+
+export const customizations = {
+  actions: [...prompts.actions],
+  views: [...prompts.views],
+};
