+++ server/slack/features/customizations/prompts/actions/edit-prompt.ts
@@ -1,0 +1,35 @@
+import type {
+  AllMiddlewareArgs,
+  BlockAction,
+  ButtonAction,
+  SlackActionMiddlewareArgs,
+} from '@slack/bolt';
+import { getUserCustomization } from '~/db/queries/customizations';
+import logger from '~/lib/logger';
+import { toLogError } from '~/utils/error';
+import { buildPromptModal } from '../view';
+
+export const name = 'home_edit_prompt';
+
+export async function execute({
+  ack,
+  body,
+  client,
+}: SlackActionMiddlewareArgs<BlockAction<ButtonAction>> &
+  AllMiddlewareArgs): Promise<void> {
+  await ack();
+  const userId = body.user.id;
+  const currentCustomization = await getUserCustomization(userId).catch(
+    (error) => {
+      logger.warn(
+        { ...toLogError(error), userId },
+        'Failed to fetch customization for modal'
+      );
+      return null;
+    }
+  );
+  await client.views.open({
+    trigger_id: body.trigger_id,
+    view: buildPromptModal(currentCustomization?.prompt ?? null),
+  });
+}
